requirements.txt
pyautogen[gemini]>=0.2.0
python-dotenv
fastapi>=0.109.0
uvicorn>=0.27.0